+++ Config/ProjectSettings.ps1
@@ -610,8 +610,6 @@ if ($Develop -or !(Get-Variable -Name CheckReadOnlyVariables2 -Scope Global -Err
 	# If there are no standard users you should set this to "Administrators" or some other group
 	# which represents non administrative users on target computer.
 	# The default value is "Users"
-	# TODO: A good portion of code handles only first group mentioned, and in some places Users and
-	# Administrators are hardcoded while DefaultGroup should be used
 	Set-Variable -Name DefaultGroup -Scope Global -Option ReadOnly -Force -Value @(
 		# Add or remove groups as needed
 		"Users"