+++ docs/CHANGELOG.md
@@ -32,6 +32,7 @@ Here is a list of changes for each of the releases.
   - Added workaround rules for `Microsoft Teams` app present in Windows 11
   - Added rule for Network List Service
   - Updated DNS rule to include port for encrypted DNS on Windows 11
+  - Updated rule related to Windows 11 updates
 
 - Modules